Added generated item spawner to not have to deal with in scene objects :)
