+++ Assets/Editor/DebugScripts/GeneratedItemSpawnerDebug.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 54406a4ed9eca31488ca429db3a91e46
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
