+++ Assets/Scripts/Equipment/GeneratedItemSpawner.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 803916649c0368943a8573213af9e8b0
+guid: f65d3ece1968be34782a7515c402ba79
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
